requirements.txt
python-telegram-bot[job-queue]>=21.0,<22.0
notion-client>=2.2.1
APScheduler>=3.10.4
python-dotenv>=1.0.1
pytz>=2024.1
